Project completed
* implemented html structure

+ indent using tabs

* sat background color for website

* h4 as h4 - selecting not by class name

* defined font weights

* added bg, renamed favicon + fixed favicon

* set body & root background

+ body background set to be a cover
+ #root set bg color that is bg for the QR-card on larger devices than mobile

* redundant image container element

* center qr card

* styling qr code

* styling img

* justifying to center

+ rounder radius for card
+ img to center

* layout center successful

* image is now fitting in by card's layout

* enlarged card

+ set color for <p>
+ adjusted margin for p & h4
+ changed font size of <h4>
+ increased card width,

* adjusted values

+ centered <p>
+ reduced margin bottom of <h4> and increased it's size

* adjusted card

* applied shadow

* scaled QR card

I don't know why the scale is too high, so i used transform instead of trying to find "why". The font size is ok and the value units were fine.
anyway, i came to a conclusion that using scale() is the easiest and laziest way in this case - small project.

* responsive

used media queries
why removed margin of #root:
so that the default view is mobile, not desktop.